# apps/high-steak-api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: High Steak API
  description: |
    REST API for the High Steak social steak-rating app.

    ## Authorization

    Most endpoints require `Authorization: Bearer <JWT>`. Each operation lists required permission scopes in **`x-required-scopes`**. The JWT must include those values in its `scopes` claim (see login/register token).

    **Role and scopes are never returned in JSON responses** — clients must read `roles` and `scopes` from the JWT payload only.
  version: 1.0.0
  contact:
    name: High Steak
servers:
  - url: http://localhost:8080/api
    description: Local development
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Username or email already taken
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/check-username:
    get:
      tags: [Auth]
      summary: Check username availability
      description: |
        Validates username format (letters, digits, `_`, `-`; must start with a letter)
        and reports whether the handle is already taken. Public — no auth required.
      operationId: checkUsername
      parameters:
        - name: username
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Username'
      responses:
        '200':
          description: Availability result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
  /auth/check-email:
    get:
      tags: [Auth]
      summary: Check email availability
      description: Validates email format and reports whether it is already registered.
      operationId: checkEmail
      parameters:
        - name: email
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Email'
      responses:
        '200':
          description: Availability result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getMe
      x-required-scopes: []
      description: Valid JWT required. No specific scope. Role/scopes are not returned — read from token.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Auth]
      summary: Update current user profile
      operationId: updateMe
      x-required-scopes: []
      description: Valid JWT required. No specific scope. Returns refreshed JWT; role/scopes are not in the response body.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateProfileMultipartForm'
      responses:
        '200':
          description: Updated profile with refreshed JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /posts:
    get:
      tags: [Posts]
      summary: Authenticated feed (non-hidden posts)
      operationId: getFeed
      x-required-scopes:
        - posts:read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Posts]
      summary: Create a post
      operationId: createPost
      x-required-scopes:
        - posts:write
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreatePostForm'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /posts/review-tags:
    get:
      tags: [Posts]
      summary: List predefined quick-review tags
      operationId: getReviewTags
      x-required-scopes:
        - posts:read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Positive and negative tag catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTagCatalog'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /posts/mine:
    get:
      tags: [Posts]
      summary: List current user's posts
      operationId: getMyPosts
      x-required-scopes:
        - posts:read:own
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /posts/hidden:
    get:
      tags: [Posts]
      summary: List hidden posts (moderators)
      operationId: getHiddenPosts
      x-required-scopes:
        - posts:moderate
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Hidden posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /posts/{id}:
    get:
      tags: [Posts]
      summary: Get a single post
      description: |
        Returns a post if the caller may view it. Followers-only posts return 404 for
        non-followers (and non-moderators) to avoid leaking existence.
      operationId: getPost
      x-required-scopes:
        - posts:read
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Posts]
      summary: Update own post
      operationId: updatePost
      x-required-scopes:
        - posts:write
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdatePostForm'
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Posts]
      summary: Delete a post
      description: Post owner needs posts:delete:own; moderators/admins may use posts:delete:any.
      operationId: deletePost
      x-required-scopes:
        - posts:delete:own
        - posts:delete:any
      x-required-scopes-one-of: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '204':
          description: Post deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /posts/{id}/hide:
    patch:
      tags: [Posts]
      summary: Hide a post (moderators)
      operationId: hidePost
      x-required-scopes:
        - posts:moderate
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Post hidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /posts/{id}/comments:
    get:
      tags: [Posts]
      summary: List comments on a post
      operationId: listPostComments
      x-required-scopes:
        - posts:read
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Posts]
      summary: Add a comment to a post
      operationId: addPostComment
      x-required-scopes:
        - comments:write
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /posts/following:
    get:
      tags: [Posts]
      summary: Feed from subscribed users
      operationId: getFollowingFeed
      x-required-scopes:
        - subscriptions:read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Posts from followed users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT from `POST /auth/login` or `POST /auth/register`.

        Decode the payload to read authorization data (never returned in JSON bodies):
        - `roles` — array with one role name (e.g. `USER`, `MODERATOR`, `ADMIN`)
        - `scopes` — permission strings checked by the API (see `x-required-scopes` on each operation)

        Known scopes:
        - `posts:read`, `posts:read:own`, `posts:write`, `posts:delete:own`, `posts:delete:any`, `posts:moderate`
        - `comments:write`
        - `users:discover`, `users:read`, `users:manage`
        - `subscriptions:read`, `subscriptions:write`
...
  parameters:
    PostId:
      name: id
      in: path
      required: true
      description: Steak post UUID.
      schema:
        type: string
        format: uuid
      example: b2c3d4e5-f6a7-8901-bcde-f12345678901
  schemas:
    Username:
      type: string
      minLength: 3
      maxLength: 50
      pattern: '^[a-zA-Z][a-zA-Z0-9_-]*$'
      description: |
        Unique login handle (`users.username`). Letters, digits, underscores, and hyphens only;
        must start with a letter (not a number).
      example: steak_lover
    AvailabilityResponse:
      type: object
      required: [available, message]
      properties:
        available:
          type: boolean
          description: True when the value is valid and not already taken.
        message:
          type: string
          description: Human-readable validation or availability message.
          example: Username is available
    Email:
      type: string
      format: email
      maxLength: 255
      description: Account email stored in `users.email` (VARCHAR 255).
      example: chef@example.com
    Password:
      type: string
      format: password
      minLength: 8
      maxLength: 100
      description: Plain-text password for login/register (stored hashed server-side).
      example: s3cur3Pass!
    DisplayName:
      type: string
      minLength: 2
      maxLength: 100
      description: Public name shown on posts and profiles (`users.display_name`).
      example: Hossam Kandel
    PostTitle:
      type: string
      minLength: 1
      maxLength: 120
      description: Short headline for the steak post (`steak_posts.title`).
      example: Perfect medium-rare ribeye
    PostComment:
      type: string
      maxLength: 65535
      description: Optional long-form caption (`steak_posts.comment`, TEXT).
      example: 28-day dry-aged ribeye, reverse seared, finished with rosemary butter.
    RestaurantName:
      type: string
      maxLength: 120
      description: Optional venue name (`steak_posts.restaurant_name`).
      example: The Prime Cut
    RestaurantLocation:
      type: string
      maxLength: 255
      description: Optional city or address hint (`steak_posts.restaurant_location`).
      example: Austin, TX
    SteakRating:
      type: integer
      format: int32
      minimum: 1
      maximum: 5
      description: Overall score from 1 (poor) to 5 (exceptional).
      example: 5
    PostVisibility:
      type: string
      enum: [PUBLIC, FOLLOWERS_ONLY]
      default: PUBLIC
      description: |
        Who can see the post. `PUBLIC` appears in the everyone feed.
        `FOLLOWERS_ONLY` is visible to the author, their followers, and moderators only.
      example: PUBLIC
    ImageUpload:
      type: string
      format: binary
      description: Image file (JPEG, PNG, WebP, etc.). Each file must be ≤ 1 MB.
    TagIdList:
      type: array
      maxItems: 12
      items:
        type: string
        format: uuid
      description: Quick-review tag IDs from `GET /posts/review-tags` (max 12).
    RegisterRequest:
      type: object
      required: [username, email, password, displayName]
      properties:
        username:
          $ref: '#/components/schemas/Username'
        email:
          $ref: '#/components/schemas/Email'
        password:
          $ref: '#/components/schemas/Password'
        displayName:
          $ref: '#/components/schemas/DisplayName'
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          maxLength: 50
          description: Registered username.
          example: steaklover42
        password:
          $ref: '#/components/schemas/Password'
    AuthResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: "JWT access token used as `Authorization: Bearer <token>`."
          example: eyJhbGciOiJIUzI1NiJ9...
    UpdateProfileRequest:
      type: object
      description: At least one field should be provided when updating profile.
      properties:
        displayName:
          $ref: '#/components/schemas/DisplayName'
        email:
          $ref: '#/components/schemas/Email'
    UpdateProfileMultipartForm:
      type: object
      description: Multipart alternative to JSON profile update. All fields optional.
      properties:
        displayName:
          $ref: '#/components/schemas/DisplayName'
        email:
          $ref: '#/components/schemas/Email'
        avatar:
          $ref: '#/components/schemas/ImageUpload'
    CreatePostForm:
      type: object
      required: [title, rating, images]
      properties:
        title:
          $ref: '#/components/schemas/PostTitle'
        comment:
          $ref: '#/components/schemas/PostComment'
        rating:
          $ref: '#/components/schemas/SteakRating'
        restaurantName:
          $ref: '#/components/schemas/RestaurantName'
        restaurantLocation:
          $ref: '#/components/schemas/RestaurantLocation'
        images:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ImageUpload'
          description: One or more steak photos. At least one required; each ≤ 1 MB.
        tagIds:
          $ref: '#/components/schemas/TagIdList'
        visibility:
          $ref: '#/components/schemas/PostVisibility'
    UpdatePostForm:
      type: object
      required: [title, rating]
      properties:
        title:
          $ref: '#/components/schemas/PostTitle'
        comment:
          $ref: '#/components/schemas/PostComment'
        rating:
          $ref: '#/components/schemas/SteakRating'
        restaurantName:
          $ref: '#/components/schemas/RestaurantName'
        restaurantLocation:
          $ref: '#/components/schemas/RestaurantLocation'
        keepImageUrls:
          type: array
          items:
            type: string
            maxLength: 512
          description: Existing image URLs to retain, in display order.
          example: [/uploads/abc123.jpg]
        images:
          type: array
          items:
            $ref: '#/components/schemas/ImageUpload'
          description: New photos to append. Each ≤ 1 MB. Combined with kept URLs, at least one image must remain.
        tagIds:
          $ref: '#/components/schemas/TagIdList'
        visibility:
          $ref: '#/components/schemas/PostVisibility'
    UpdateProfileResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: Refreshed JWT reflecting updated profile claims.
        user:
          $ref: '#/components/schemas/UserSummary'
    UserSummary:
      type: object
      required: [id, username, email, displayName]
      description: |
        Public account profile. Does **not** include role or scopes — read those from the JWT `roles` and `scopes` claims.
      properties:
        id:
          type: string
          format: uuid
          description: User UUID.
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        username:
          $ref: '#/components/schemas/Username'
        email:
          $ref: '#/components/schemas/Email'
        displayName:
          $ref: '#/components/schemas/DisplayName'
        avatarUrl:
          type: string
          nullable: true
          maxLength: 512
          description: Relative URL to profile photo, or null if unset.
          example: /uploads/avatar-abc.jpg
...
    AuthorSummary:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: string
          format: uuid
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        displayName:
          $ref: '#/components/schemas/DisplayName'
    PostResponse:
      type: object
      required: [id, title, rating, imageUrls, createdAt, hidden, visibility, author, tags]
      properties:
        id:
          type: string
          format: uuid
          example: b2c3d4e5-f6a7-8901-bcde-f12345678901
        title:
          $ref: '#/components/schemas/PostTitle'
        comment:
          allOf:
            - $ref: '#/components/schemas/PostComment'
          nullable: true
        rating:
          $ref: '#/components/schemas/SteakRating'
        imageUrls:
          type: array
          minItems: 1
          items:
            type: string
            maxLength: 512
          description: Ordered photo URLs for the post.
          example: [/uploads/abc123.jpg, /uploads/def456.jpg]
        restaurantName:
          allOf:
            - $ref: '#/components/schemas/RestaurantName'
          nullable: true
        restaurantLocation:
          allOf:
            - $ref: '#/components/schemas/RestaurantLocation'
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: When the post was published (ISO-8601 UTC).
          example: '2026-05-30T18:30:00Z'
        hidden:
          type: boolean
          description: True if moderators hid the post from public feeds.
          example: false
        visibility:
          $ref: '#/components/schemas/PostVisibility'
        author:
          $ref: '#/components/schemas/AuthorSummary'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/ReviewTagSummary'
    ReviewTagSummary:
      type: object
      required: [id, label, sentiment]
      properties:
        id:
          type: string
          format: uuid
          example: c3d4e5f6-a7b8-9012-cdef-123456789012
        label:
          type: string
          maxLength: 80
          description: Human-readable tag label.
          example: Juicy
        sentiment:
          type: string
          enum: [POSITIVE, NEGATIVE]
          example: POSITIVE
    ReviewTagCatalog:
      type: object
      required: [positive, negative]
      properties:
        positive:
          type: array
          items:
            $ref: '#/components/schemas/ReviewTagSummary'
        negative:
          type: array
          items:
            $ref: '#/components/schemas/ReviewTagSummary'
    CommentResponse:
      type: object
      required: [id, body, createdAt, author]
      properties:
        id:
          type: string
          format: uuid
          example: d4e5f6a7-b8c9-0123-def0-234567890123
        body:
          type: string
          maxLength: 2000
          description: Comment text (`post_comments.body`, TEXT capped at 2000 in API).
          example: That crust looks incredible!
        createdAt:
          type: string
          format: date-time
          example: '2026-05-30T19:00:00Z'
        author:
          $ref: '#/components/schemas/AuthorSummary'
    CreateCommentRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 2000
          description: Comment text (1–2000 characters).
          example: What temperature did you pull it at?
    ApiError:
      type: object
      required: [timestamp, status, error, message]
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2026-05-30T19:00:00Z'
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Title must be at most 120 characters
        fieldErrors:
          type: object
          additionalProperties:
            type: string
          nullable: true
          example:
            title: must not be blank
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'